Changed eigensolver interface as c_int overspills

+++ prog/dftb+/lib_math/eigensolver.F90
@@ -18,6 +18,7 @@ module eigensolver
   use lapack
 #:if WITH_GPU
   use magma
+  use iso_c_binding, only :  c_int
 #:endif
 
   !implicit none
@@ -2065,7 +2066,7 @@ contains
   subroutine ${VTYPE}$_magma_${VPREC}$${NAME}$(ngpus, a, b, w, uplo, jobz, itype)
 
     !> Number of GPUs to use
-    integer, intent(in) :: ngpus
+    integer(c_int), intent(in) :: ngpus
 
     !> contains the matrix for the solver, returns eigenvectors if requested (matrix always
     !> overwritten on return anyway)
